+++ tezos/client/benches/bench_apply_first_three_blocks.rs
@@ -8,7 +8,7 @@ use test::Bencher;
 
 use crypto::hash::ChainId;
 use tezos_api::environment::{
-    get_empty_operation_list_list_hash, TezosEnvironmentConfiguration, TEZOS_ENV,
+    default_networks, get_empty_operation_list_list_hash, TezosEnvironmentConfiguration,
 };
 use tezos_api::ffi::{
     ApplyBlockRequest, InitProtocolContextResult, TezosContextConfiguration,
@@ -138,7 +138,8 @@ fn apply_first_three_blocks(
 }
 
 fn init_test_protocol_context(dir_name: &str) -> (ChainId, BlockHeader, InitProtocolContextResult) {
-    let tezos_env: &TezosEnvironmentConfiguration = TEZOS_ENV
+    let default_networks = default_networks();
+    let tezos_env: &TezosEnvironmentConfiguration = default_networks
         .get(&test_data::TEZOS_NETWORK)
         .expect("no tezos environment configured");