+++ tezos/client/tests/prevalidator_mempool_test.rs
@@ -5,7 +5,7 @@ use serial_test::serial;
 
 use crypto::hash::{ChainId, ProtocolHash};
 use tezos_api::environment::{
-    get_empty_operation_list_list_hash, TezosEnvironmentConfiguration, TEZOS_ENV,
+    default_networks, get_empty_operation_list_list_hash, TezosEnvironmentConfiguration,
 };
 use tezos_api::ffi::{
     ApplyBlockRequest, BeginConstructionRequest, InitProtocolContextResult,
@@ -37,7 +37,8 @@ fn init_test_protocol_context(
     ProtocolHash,
     InitProtocolContextResult,
 ) {
-    let tezos_env: &TezosEnvironmentConfiguration = TEZOS_ENV
+    let default_networks = default_networks();
+    let tezos_env: &TezosEnvironmentConfiguration = default_networks
         .get(&test_data::TEZOS_NETWORK)
         .expect("no tezos environment configured");